Move 'RsaKeyPair' implementation to its own module.
